chunking/requirements.txt
google-cloud-storage
markdownify
langchain-text-splitters
langchain
beautifulsoup4
kfp
google-cloud-aiplatform
